updatd the activity feed with ai summary
