+++ src/components/ActivityFeed/PostList.tsx
@@ -3,7 +3,7 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 import React from 'react';
 import PostItem from './PostItem';
-import { PostListProps } from './utils/types';
+import { PostListProps } from './types/types';
 
 const PostList: React.FC<PostListProps> = ({ postData }) => {
 	return (
