+++ src/util/nextApiClientFetch.ts
@@ -8,22 +8,33 @@ import getNetwork from './getNetwork';
 import messages from './messages';
 import reAuthClient from './reAuthClient';
 
-async function nextApiClientFetch<T>(url: string, data?: { [key: string]: any }, method?: 'GET' | 'POST'): Promise<{ data?: T; error?: string }> {
+async function nextApiClientFetch<T>(url: string, data?: { [key: string]: any; aiSiummaryKey?: boolean }, method?: 'GET' | 'POST'): Promise<{ data?: T; error?: string }> {
 	const network = getNetwork();
 
 	const currentURL = new URL(window.location.href);
 	const token = currentURL.searchParams.get('token') || (await reAuthClient()) || getLocalStorageToken();
 
+	const Payload: {
+		Authorization: string;
+		'Content-Type': string;
+		'x-api-key': string;
+		'x-network': string;
+		'x-ai-summary-key'?: string;
+	} = {
+		Authorization: `Bearer ${token}`,
+		'Content-Type': 'application/json',
+		'x-api-key': process.env.NEXT_PUBLIC_POLKASSEMBLY_API_KEY ?? '',
+		'x-network': network
+	};
+
+	if (data?.aiSiummaryKey) {
+		Payload['x-ai-summary-key'] = process.env.API_SUMMARY_PROVIDER;
+	}
 	try {
 		const response = await fetch(`${window.location.origin}/${url}`, {
 			body: JSON.stringify(data),
 			credentials: 'include',
-			headers: {
-				Authorization: `Bearer ${token}`,
-				'Content-Type': 'application/json',
-				'x-api-key': process.env.NEXT_PUBLIC_POLKASSEMBLY_API_KEY ?? '',
-				'x-network': network
-			},
+			headers: Payload,
 			method: method ?? 'POST'
 		});
